refactor: server 패키지를 data 패키지의 하위 패키지로 이동
