+++ app/src/main/java/woowacourse/shopping/data/server/MockWebServer.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.server
+package woowacourse.shopping.data.server
 
 import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
