+++ app/src/main/java/woowacourse/shopping/data/server/ProductRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.server
+package woowacourse.shopping.data.server
 
 import okhttp3.Call
 import okhttp3.Callback
