+++ app/src/main/java/woowacourse/shopping/data/server/Products.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.server
+package woowacourse.shopping.data.server
 
 val productsPage1 = """
     |[
